Move Dealer class into single file
